extract schemaoptions
